requirements.txt
# =============================================================================
# Chakula Poa Backend - Python Dependencies
# =============================================================================
# Django & REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9    # PostgreSQL adapter
dj-database-url>=2.1.0    # Parse DATABASE_URL

# Environment & Configuration
python-decouple>=3.8      # Environment variables

# Image Processing & QR Codes
Pillow>=10.1.0            # Image processing
qrcode>=7.4.2             # QR code generation

# Production Server
gunicorn>=21.2.0          # WSGI HTTP Server
whitenoise>=6.6.0         # Static file serving

# Optional: Development tools
# pip install -r requirements-dev.txt for these
